Separate PSR11 and PHPDI kernels
