+++ src/KernelWithPhpDiContainer.php
@@ -10,9 +10,6 @@
 namespace DI\Bridge\Symfony;
 
 use Psr\Container\ContainerInterface;
-use Symfony\Component\Debug\DebugClassLoader;
-use Symfony\Component\DependencyInjection\Compiler\CheckExceptionOnInvalidReferenceBehaviorPass;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
  * Customization of Symfony's kernel to setup PHP-DI.
@@ -21,18 +18,12 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
-abstract class KernelWithPhpDiContainer extends \Symfony\Component\HttpKernel\Kernel
+abstract class KernelWithPhpDiContainer extends KernelWithPsr11Container
 {
     /**
      * @var ContainerInterface
      */
     private $phpdiContainer;
-
-    public function __construct($environment, $debug)
-    {
-        parent::__construct($environment, $debug);
-        $this->disableDebugClassLoader();
-    }
 
     /**
      * Implement this method to configure PHP-DI.
@@ -41,68 +32,10 @@ abstract class KernelWithPhpDiContainer extends \Symfony\Component\HttpKernel\Ke
      */
     abstract protected function buildPHPDIContainer(\DI\ContainerBuilder $builder);
 
-    protected function getContainerBaseClass()
-    {
-        return SymfonyContainerBridge::class;
-    }
-
-    protected function buildContainer()
-    {
-        $containerBuilder = parent::buildContainer();
-
-        $this->removeInvalidReferenceBehaviorPass($containerBuilder);
-
-        return $containerBuilder;
-    }
-
-    protected function initializeContainer()
-    {
-        parent::initializeContainer();
-
-        /** @var SymfonyContainerBridge $rootContainer */
-        $rootContainer = $this->getContainer();
-
-        $rootContainer->setFallbackContainer($this->getPHPDIContainer());
-    }
-
-    /**
-     * Remove the CheckExceptionOnInvalidReferenceBehaviorPass because
-     * it was not looking into PHP-DI's entries and thus throwing exceptions.
-     *
-     * @todo Replace it by an alternative that can search into PHP-DI too
-     *       Problem: PHP-DI is not initialized when Symfony's container is compiled, because
-     *       it depends on Symfony's container for fallback (cycle…)
-     *
-     * @param ContainerBuilder $container
-     */
-    private function removeInvalidReferenceBehaviorPass(ContainerBuilder $container)
-    {
-        $passConfig = $container->getCompilerPassConfig();
-        $compilationPasses = $passConfig->getRemovingPasses();
-
-        foreach ($compilationPasses as $i => $pass) {
-            if ($pass instanceof CheckExceptionOnInvalidReferenceBehaviorPass) {
-                unset($compilationPasses[$i]);
-                break;
-            }
-        }
-
-        $passConfig->setRemovingPasses($compilationPasses);
-    }
-
-    private function disableDebugClassLoader()
-    {
-        if (!class_exists(DebugClassLoader::class)) {
-            return;
-        }
-
-        DebugClassLoader::disable();
-    }
-
     /**
      * @return ContainerInterface
      */
-    protected function getPHPDIContainer()
+    protected function getFallbackContainer()
     {
         if ($this->phpdiContainer === null) {
             $builder = new \DI\ContainerBuilder();
